requirements.txt
flask
pinecone
sentence-transformers
langchain
PyPDF2
python-dotenv